+++ src/Renderer/mediainfo.h
@@ -163,6 +163,7 @@ public:
     void setCrop(int top, int bottom, int left, int right, int id = -1, bool silent = false);
     void setCrop(int width, int height, int id = -1, bool silent = false);
     void setCropUseSize(bool useSize, int id = -1, bool silent = false);
+    void setLut(QString lut, int id = -1, bool silent = false);
     //audio
     void setSamplingRate(int value, int id = -1, bool silent = false);
     void setChannels(QString value, int id = -1, bool silent = false);