+++ src/UI/Blocks/blocklut.h
@@ -1,0 +1,24 @@
+#ifndef BLOCKLUT_H
+#define BLOCKLUT_H
+
+#include "ui_blocklut.h"
+#include "UI/Blocks/blockcontentwidget.h"
+#include "Renderer/cachemanager.h"
+
+#include <QFileDialog>
+
+class BlockLut : public BlockContentWidget,  private Ui::BlockLut
+{
+    Q_OBJECT
+
+public:
+    explicit BlockLut(MediaInfo *mediaInfo, QWidget *parent = nullptr);
+
+public slots:
+    void activate( bool blockEnabled );
+    void update();
+private slots:
+    void on_lutBox_currentIndexChanged(int index);
+};
+
+#endif // BLOCKLUT_H