+++ src/global.cpp
@@ -2,4 +2,4 @@
 
 PresetManager *presetManager = new PresetManager();
 FFmpeg *ffmpeg = new FFmpeg();
-CacheManager *cacheManager = new CacheManager();
+CacheManager *cacheManager = CacheManager::instance();